Rename nanopb generated headers from foo.pb.h to foo.nanopb.h
This avoids a collision with the libprotobuf generated files.
