+++ Firestore/Protos/nanopb/firestore/local/target.nanopb.h
@@ -15,15 +15,15 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Mon Apr  9 15:08:47 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
 #ifndef PB_FIRESTORE_CLIENT_TARGET_PB_H_INCLUDED
 #define PB_FIRESTORE_CLIENT_TARGET_PB_H_INCLUDED
 #include <pb.h>
 
-#include "google/firestore/v1beta1/firestore.pb.h"
+#include "google/firestore/v1beta1/firestore.nanopb.h"
 
-#include "google/protobuf/timestamp.pb.h"
+#include "google/protobuf/timestamp.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
