+++ Firestore/Protos/nanopb/firestore/local/target.pb.c
@@ -15,9 +15,9 @@
  */
 
 /* Automatically generated nanopb constant definitions */
-/* Generated by nanopb-0.3.8 at Mon Apr  9 15:08:47 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#include "target.pb.h"
+#include "target.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
