+++ Firestore/Protos/nanopb/google/api/annotations.nanopb.h
@@ -14,21 +14,31 @@
  * limitations under the License.
  */
 
-/* Automatically generated nanopb constant definitions */
+/* Automatically generated nanopb header */
 /* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#include "empty.nanopb.h"
+#ifndef PB_GOOGLE_API_ANNOTATIONS_PB_H_INCLUDED
+#define PB_GOOGLE_API_ANNOTATIONS_PB_H_INCLUDED
+#include <pb.h>
+
+#include "google/api/http.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
 #error Regenerate this file with the current version of nanopb generator.
 #endif
 
+#ifdef __cplusplus
+extern "C" {
+#endif
 
+/* Extensions */
+/* Extension field google_api_http was skipped because only "optional"
+   type of extension fields is currently supported. */
 
-const pb_field_t google_protobuf_Empty_fields[1] = {
-    PB_LAST_FIELD
-};
-
-
+#ifdef __cplusplus
+} /* extern "C" */
+#endif
 /* @@protoc_insertion_point(eof) */
+
+#endif
