+++ Firestore/Protos/nanopb/google/api/annotations.pb.c
@@ -15,9 +15,9 @@
  */
 
 /* Automatically generated nanopb constant definitions */
-/* Generated by nanopb-0.3.8 at Fri Feb  2 17:48:02 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#include "annotations.pb.h"
+#include "annotations.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
