+++ Firestore/Protos/nanopb/google/firestore/v1beta1/common.nanopb.h
@@ -15,15 +15,15 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Fri Feb  2 17:48:02 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
 #ifndef PB_GOOGLE_FIRESTORE_V1BETA1_COMMON_PB_H_INCLUDED
 #define PB_GOOGLE_FIRESTORE_V1BETA1_COMMON_PB_H_INCLUDED
 #include <pb.h>
 
-#include "google/api/annotations.pb.h"
+#include "google/api/annotations.nanopb.h"
 
-#include "google/protobuf/timestamp.pb.h"
+#include "google/protobuf/timestamp.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
