+++ Firestore/Protos/nanopb/google/firestore/v1beta1/query.nanopb.h
@@ -15,17 +15,17 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Thu Apr 12 07:27:15 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
 #ifndef PB_GOOGLE_FIRESTORE_V1BETA1_QUERY_PB_H_INCLUDED
 #define PB_GOOGLE_FIRESTORE_V1BETA1_QUERY_PB_H_INCLUDED
 #include <pb.h>
 
-#include "google/api/annotations.pb.h"
+#include "google/api/annotations.nanopb.h"
 
-#include "google/firestore/v1beta1/document.pb.h"
+#include "google/firestore/v1beta1/document.nanopb.h"
 
-#include "google/protobuf/wrappers.pb.h"
+#include "google/protobuf/wrappers.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
