+++ Firestore/Protos/nanopb/google/protobuf/any.nanopb.h
@@ -15,10 +15,10 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Mon Feb 12 11:03:06 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#ifndef PB_GOOGLE_PROTOBUF_EMPTY_PB_H_INCLUDED
-#define PB_GOOGLE_PROTOBUF_EMPTY_PB_H_INCLUDED
+#ifndef PB_GOOGLE_PROTOBUF_ANY_PB_H_INCLUDED
+#define PB_GOOGLE_PROTOBUF_ANY_PB_H_INCLUDED
 #include <pb.h>
 
 /* @@protoc_insertion_point(includes) */
@@ -31,29 +31,32 @@ extern "C" {
 #endif
 
 /* Struct definitions */
-typedef struct _google_protobuf_Empty {
-    char dummy_field;
-/* @@protoc_insertion_point(struct:google_protobuf_Empty) */
-} google_protobuf_Empty;
+typedef struct _google_protobuf_Any {
+    pb_callback_t type_url;
+    pb_callback_t value;
+/* @@protoc_insertion_point(struct:google_protobuf_Any) */
+} google_protobuf_Any;
 
 /* Default values for struct fields */
 
 /* Initializer values for message structs */
-#define google_protobuf_Empty_init_default       {0}
-#define google_protobuf_Empty_init_zero          {0}
+#define google_protobuf_Any_init_default         {{{NULL}, NULL}, {{NULL}, NULL}}
+#define google_protobuf_Any_init_zero            {{{NULL}, NULL}, {{NULL}, NULL}}
 
 /* Field tags (for use in manual encoding/decoding) */
+#define google_protobuf_Any_type_url_tag         1
+#define google_protobuf_Any_value_tag            2
 
 /* Struct field encoding specification for nanopb */
-extern const pb_field_t google_protobuf_Empty_fields[1];
+extern const pb_field_t google_protobuf_Any_fields[3];
 
 /* Maximum encoded size of messages (where known) */
-#define google_protobuf_Empty_size               0
+/* google_protobuf_Any_size depends on runtime parameters */
 
 /* Message IDs (where set with "msgid" option) */
 #ifdef PB_MSGID
 
-#define EMPTY_MESSAGES \
+#define ANY_MESSAGES \
 
 
 #endif
