+++ Firestore/Protos/nanopb/google/protobuf/empty.nanopb.h
@@ -15,13 +15,11 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Fri Feb  2 17:48:02 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#ifndef PB_GOOGLE_API_ANNOTATIONS_PB_H_INCLUDED
-#define PB_GOOGLE_API_ANNOTATIONS_PB_H_INCLUDED
+#ifndef PB_GOOGLE_PROTOBUF_EMPTY_PB_H_INCLUDED
+#define PB_GOOGLE_PROTOBUF_EMPTY_PB_H_INCLUDED
 #include <pb.h>
-
-#include "google/api/http.pb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
@@ -32,9 +30,33 @@
 extern "C" {
 #endif
 
-/* Extensions */
-/* Extension field google_api_http was skipped because only "optional"
-   type of extension fields is currently supported. */
+/* Struct definitions */
+typedef struct _google_protobuf_Empty {
+    char dummy_field;
+/* @@protoc_insertion_point(struct:google_protobuf_Empty) */
+} google_protobuf_Empty;
+
+/* Default values for struct fields */
+
+/* Initializer values for message structs */
+#define google_protobuf_Empty_init_default       {0}
+#define google_protobuf_Empty_init_zero          {0}
+
+/* Field tags (for use in manual encoding/decoding) */
+
+/* Struct field encoding specification for nanopb */
+extern const pb_field_t google_protobuf_Empty_fields[1];
+
+/* Maximum encoded size of messages (where known) */
+#define google_protobuf_Empty_size               0
+
+/* Message IDs (where set with "msgid" option) */
+#ifdef PB_MSGID
+
+#define EMPTY_MESSAGES \
+
+
+#endif
 
 #ifdef __cplusplus
 } /* extern "C" */
