+++ Firestore/Protos/nanopb/google/rpc/status.nanopb.h
@@ -15,11 +15,13 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Fri Feb  2 17:48:02 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#ifndef PB_GOOGLE_PROTOBUF_TIMESTAMP_PB_H_INCLUDED
-#define PB_GOOGLE_PROTOBUF_TIMESTAMP_PB_H_INCLUDED
+#ifndef PB_GOOGLE_RPC_STATUS_PB_H_INCLUDED
+#define PB_GOOGLE_RPC_STATUS_PB_H_INCLUDED
 #include <pb.h>
+
+#include "google/protobuf/any.nanopb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
@@ -31,32 +33,34 @@ extern "C" {
 #endif
 
 /* Struct definitions */
-typedef struct _google_protobuf_Timestamp {
-    int64_t seconds;
-    int32_t nanos;
-/* @@protoc_insertion_point(struct:google_protobuf_Timestamp) */
-} google_protobuf_Timestamp;
+typedef struct _google_rpc_Status {
+    int32_t code;
+    pb_callback_t message;
+    pb_callback_t details;
+/* @@protoc_insertion_point(struct:google_rpc_Status) */
+} google_rpc_Status;
 
 /* Default values for struct fields */
 
 /* Initializer values for message structs */
-#define google_protobuf_Timestamp_init_default   {0, 0}
-#define google_protobuf_Timestamp_init_zero      {0, 0}
+#define google_rpc_Status_init_default           {0, {{NULL}, NULL}, {{NULL}, NULL}}
+#define google_rpc_Status_init_zero              {0, {{NULL}, NULL}, {{NULL}, NULL}}
 
 /* Field tags (for use in manual encoding/decoding) */
-#define google_protobuf_Timestamp_seconds_tag    1
-#define google_protobuf_Timestamp_nanos_tag      2
+#define google_rpc_Status_code_tag               1
+#define google_rpc_Status_message_tag            2
+#define google_rpc_Status_details_tag            3
 
 /* Struct field encoding specification for nanopb */
-extern const pb_field_t google_protobuf_Timestamp_fields[3];
+extern const pb_field_t google_rpc_Status_fields[4];
 
 /* Maximum encoded size of messages (where known) */
-#define google_protobuf_Timestamp_size           22
+/* google_rpc_Status_size depends on runtime parameters */
 
 /* Message IDs (where set with "msgid" option) */
 #ifdef PB_MSGID
 
-#define TIMESTAMP_MESSAGES \
+#define STATUS_MESSAGES \
 
 
 #endif
