+++ Firestore/Protos/nanopb/google/type/latlng.nanopb.h
@@ -15,13 +15,11 @@
  */
 
 /* Automatically generated nanopb header */
-/* Generated by nanopb-0.3.8 at Fri Feb  2 17:48:02 2018. */
+/* Generated by nanopb-0.3.8 at Fri Jun  1 18:36:44 2018. */
 
-#ifndef PB_GOOGLE_RPC_STATUS_PB_H_INCLUDED
-#define PB_GOOGLE_RPC_STATUS_PB_H_INCLUDED
+#ifndef PB_GOOGLE_TYPE_LATLNG_PB_H_INCLUDED
+#define PB_GOOGLE_TYPE_LATLNG_PB_H_INCLUDED
 #include <pb.h>
-
-#include "google/protobuf/any.pb.h"
 
 /* @@protoc_insertion_point(includes) */
 #if PB_PROTO_HEADER_VERSION != 30
@@ -33,34 +31,32 @@ extern "C" {
 #endif
 
 /* Struct definitions */
-typedef struct _google_rpc_Status {
-    int32_t code;
-    pb_callback_t message;
-    pb_callback_t details;
-/* @@protoc_insertion_point(struct:google_rpc_Status) */
-} google_rpc_Status;
+typedef struct _google_type_LatLng {
+    double latitude;
+    double longitude;
+/* @@protoc_insertion_point(struct:google_type_LatLng) */
+} google_type_LatLng;
 
 /* Default values for struct fields */
 
 /* Initializer values for message structs */
-#define google_rpc_Status_init_default           {0, {{NULL}, NULL}, {{NULL}, NULL}}
-#define google_rpc_Status_init_zero              {0, {{NULL}, NULL}, {{NULL}, NULL}}
+#define google_type_LatLng_init_default          {0, 0}
+#define google_type_LatLng_init_zero             {0, 0}
 
 /* Field tags (for use in manual encoding/decoding) */
-#define google_rpc_Status_code_tag               1
-#define google_rpc_Status_message_tag            2
-#define google_rpc_Status_details_tag            3
+#define google_type_LatLng_latitude_tag          1
+#define google_type_LatLng_longitude_tag         2
 
 /* Struct field encoding specification for nanopb */
-extern const pb_field_t google_rpc_Status_fields[4];
+extern const pb_field_t google_type_LatLng_fields[3];
 
 /* Maximum encoded size of messages (where known) */
-/* google_rpc_Status_size depends on runtime parameters */
+#define google_type_LatLng_size                  18
 
 /* Message IDs (where set with "msgid" option) */
 #ifdef PB_MSGID
 
-#define STATUS_MESSAGES \
+#define LATLNG_MESSAGES \
 
 
 #endif
